STY: move index tests
